feat: new script_error function and optional exit
The error function is renamed in script_error to emphasize the fact that it is used to specify the name of the function that calls it and the line number.
On the contrary, the new "error" function will be used when we want to display a simpler message (without the name of the function and without the line where it is called)

The exit command is now optionnal. This is usefull if we want to print some help before exiting…